simulator in its own file
